SCA-125 fix migration numbers
