+++ server/sonar-db-migration/src/it/java/org/sonar/server/platform/db/migration/version/v202502/AddProductionScopeToScaDependenciesTableIT.java
@@ -17,42 +17,37 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-
-package org.sonar.server.platform.db.migration.version.v202503;
+package org.sonar.server.platform.db.migration.version.v202502;
 
 import java.sql.SQLException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.sonar.db.MigrationDbTester;
+import org.sonar.server.platform.db.migration.step.DdlChange;
 
+import static java.sql.Types.BOOLEAN;
 import static org.sonar.db.MigrationDbTester.createForMigrationStep;
 
-class DropIndexOnScaReleasesComponentTest {
-
-  private static final String TABLE_NAME = "sca_releases";
-  private static final String COLUMN_NAME = "component_uuid";
-  private static final String INDEX_NAME = "sca_releases_comp_uuid";
+class AddProductionScopeToScaDependenciesTableIT {
+  private static final String TABLE_NAME = "sca_dependencies";
+  private static final String COLUMN_NAME = "production_scope";
 
   @RegisterExtension
-  public final MigrationDbTester db = createForMigrationStep(DropIndexOnScaReleasesComponent.class);
-  private final DropIndexOnScaReleasesComponent underTest = new DropIndexOnScaReleasesComponent(db.database());
+  public final MigrationDbTester db = createForMigrationStep(AddProductionScopeToScaDependenciesTable.class);
+  private final DdlChange underTest = new AddProductionScopeToScaDependenciesTable(db.database());
 
   @Test
-  void index_is_dropped() throws SQLException {
-    db.assertIndex(TABLE_NAME, INDEX_NAME, COLUMN_NAME);
-
+  void execute_shouldAddColumn() throws SQLException {
+    db.assertColumnDoesNotExist(TABLE_NAME, COLUMN_NAME);
     underTest.execute();
-
-    db.assertIndexDoesNotExist(TABLE_NAME, COLUMN_NAME);
+    db.assertColumnDefinition(TABLE_NAME, COLUMN_NAME, BOOLEAN, null, false);
   }
 
   @Test
-  void migration_is_reentrant() throws SQLException {
-    db.assertIndex(TABLE_NAME, INDEX_NAME, COLUMN_NAME);
-
+  void execute_shouldBeReentrant() throws SQLException {
+    db.assertColumnDoesNotExist(TABLE_NAME, COLUMN_NAME);
     underTest.execute();
     underTest.execute();
-
-    db.assertIndexDoesNotExist(TABLE_NAME, COLUMN_NAME);
+    db.assertColumnDefinition(TABLE_NAME, COLUMN_NAME, BOOLEAN, null, false);
   }
 }
