+++ server/sonar-db-migration/src/main/java/org/sonar/server/platform/db/migration/MigrationConfigurationModule.java
@@ -39,7 +39,6 @@ import org.sonar.server.platform.db.migration.version.v107.DbVersion107;
 import org.sonar.server.platform.db.migration.version.v108.DbVersion108;
 import org.sonar.server.platform.db.migration.version.v202501.DbVersion202501;
 import org.sonar.server.platform.db.migration.version.v202502.DbVersion202502;
-import org.sonar.server.platform.db.migration.version.v202503.DbVersion202503;
 
 public class MigrationConfigurationModule extends Module {
   @Override
@@ -59,7 +58,6 @@ public class MigrationConfigurationModule extends Module {
       DbVersion108.class,
       DbVersion202501.class,
       DbVersion202502.class,
-      DbVersion202503.class,
 
       // migration steps
       MigrationStepRegistryImpl.class,
